Feat: CRUD candidato

+++ frontend/src/app/candidato/page.js
@@ -1,10 +1,9 @@
 'use client';
 import { useRouter } from 'next/navigation';
 import { useUser } from '../_stores/useUser';
-import React, { useEffect } from 'react';
 import { deleteCandidato } from './api';
 import Swal from 'sweetalert2';
-import { getUserByToken } from '../../auth/api';
+import AuthComponent from '../components/AuthComponent';
 
 function CandidatoPage() {
   const router = useRouter();
@@ -12,43 +11,37 @@ function CandidatoPage() {
     user: state.user,
     setUser: state.setUser,
   }));
-  console.log(user);
-
-  useEffect(() => {
-    const token = localStorage.getItem('token');
-
-    if (!token) {
-      router.push('/login');
-    } else {
-      getUserByToken()
-        .then((res) => {
-          setUser(res.data);
-          if (res?.data?.role === 'Empresa') {
-            router.push('/empresa');
-          }
-        })
-        .catch((err) => {
-          router.push('/login');
-        });
-    }
-  }, []);
 
   const handleDelete = (id) => {
-    deleteCandidato(user.id).then((res) => {
-      Swal.fire({
-        title: 'success',
-        text: 'Deletado',
-        icon: 'success',
-      });
-      router.push('/login');
+    Swal.fire({
+      title: 'Tem certeza?',
+      text: 'Esta ação não pode ser revertida!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteCandidato(user.id).then((res) => {
+          Swal.fire({
+            title: 'Deletado',
+            text: 'Usuário deletado com sucesso',
+            icon: 'success',
+          });
+          router.push('/login');
+        });
+      }
     });
   };
 
   return (
-    <div>
-      <button onClick={() => router.push(`candidato/${user.id}/editar`)}>Editar</button>
-      <button onClick={() => handleDelete(user.id)}>Deletar</button>
-    </div>
+    <AuthComponent>
+      <div>
+        <button onClick={() => router.push(`candidato/${user.id}/editar`)}>Editar</button>
+        <button onClick={() => handleDelete(user.id)}>Deletar</button>
+      </div>
+    </AuthComponent>
   );
 }
